+++ changelog.md
@@ -9,6 +9,7 @@
 - structurizr-dsl: Adds `supportingTypes implementation-suffix <suffix>`.
 - structurizr-dsl: Fixes https://github.com/structurizr/java/issues/346 (`// comment \` joins lines).
 - structurizr-dsl: Anonymous identifiers for relationships (i.e. relationships not assigned to an identifier) are excluded from the model, and therefore also excluded from the serialised JSON.
+- structurizr-dsl: Adds a way to configure whether the DSL source is retained via a workspace property named `structurizr.dsl.source` - `true` (default) or `false`.
 
 ## 3.0.0 (19th September 2024)